Added Bella's Account

+++ CustomerData.sql
@@ -1,6 +1,6 @@
 --HELPFUL STATEMENTS
 
---delete from customer;
+--delete from customer where accountid >= 1;
 select * from customer;
 --delete from product;
 select * from product;
@@ -9,58 +9,102 @@ select * from address;
 --delete from payment;
 select * from payment;
 --delete from cart_items;
-select * from cart_items; 
+select * from cart; 
 --delete from invoice;
 select * from invoice;
 --delete from invoice_record_deleted_account;
 select * from invoice_record_deleted_account;
---delete from invoice_products
-select * from invoice_products
+select * from reviews;
+
+SELECT i.invoiceid, i.totalprice, pp.pname, c.fname || ' ' ||c.lname from invoice i join invoice_products p on i.invoiceid =p.invoiceid join product pp on pp.productid = p.productid join customer c on c.accountid= i.accountid;
+
+SELECT i.invoiceid, pp.pname from invoice i join invoice_products p on i.invoiceid =p.invoiceid join product pp on pp.productid = p.productid;
+
+SELECT i.invoiceid, i.fname || ' ' ||i.lname from invoice_record_deleted_account i;
 
 -------------------------------------------------------------------------------------------
 --BELLA ACCOUNT 
 
 --create account 
-
 INSERT into customer VALUES(seqID.nextVal, 'Isabella', 'Attisano', 'iattisan@villanova.edu', '0123456789');
---add address / payment 
+
+--Address add / delete / view
+
+--delete from address where accountid = 1 and street = '250 Spring Mill Road';
+INSERT INTO address VALUES(1, '27 W 53 Street', 'Bayonne', 'NJ', '07002');
+INSERT INTO address VALUES(1, '250 Spring Mill Road', 'Villanova', 'PA', '19085');
+
+SELECT Street || ' ' || City || ', '|| State || ', ' || Zip as Addresses FROM address WHERE accountID = 1; 
+
+--Payment add / delete / view
+
+INSERT into payment Values(1, 5, 2024, '1234567891011123', '123');
+INSERT into payment Values(1, 8, 2028, '1234123412341234', '180');
+
+select expMonth ||'/'|| expYear AS Expiration_Date, '************'||''||SUBSTR(cardnumber, 12, 4) AS Card_Number 
+FROM payment WHERE accountID = 1;
 
 --add / remove items to cart
 
+INSERT INTO cart_items VALUES(1, 1, 'medium', 1, 19.95);
+INSERT INTO cart_items VALUES(1, 4, 'large', 3, 25);
+INSERT INTO cart_items VALUES(1, 11, 'onesize', 2, 12.99);
+INSERT INTO cart_items VALUES(1, 46, 'onesize', 4, 37);
 
---see all product reviews 
+SELECT * from cart_items where cartid =1;
 
---add items to favorites / get all favorited items 
+-- total price of cart 
+select sum(c.price*c.pquantity) from cart_items c where cartid =1;
 
---get all items currently in cart / cart price 
+--see product review 
+
+select age, feet ||' foot '|| inches as height, rating ||''|| ' out of 5' as Overall_Rating, sizedesc, recommend from reviews where productid = 4;
+ 
+DELETE FROM cart_items where cartid =1 and productid = 4;
+
+UPDATE cart_items set pquantity = 1 where cartid = 1 and productid = 10;
+UPDATE cart_items set psize = 'large' where cartid = 1 and productid = 1;
 
 --place order
 
-INSERT into invoice Values(seqInvoice.nextVal, 1, 100);
-INSERT into invoice Values(seqInvoice.nextVal, 1, 500);
-INSERT into invoice_products Values(1, 1, 'large', 4);
-INSERT into invoice_products Values(2, 2, 'small', 1);
+UPDATE cart set cartid = 1 where cartid = 1;
 
+-- view invoice history 
 
---get order history 
+select * from invoice where accountid =1;
 
-select * from invoice where accountid = 1;
+--view invoice products 
+select i.invoiceid, p.pname, ip.psize, ip.pquanity, p.price 
+from invoice_products ip join invoice i on ip.invoiceid = i.invoiceid join product p on p.productid = ip.productid 
+where i.accountid = 1 
+order by i.invoiceid;
 
---add reviews 
+--add review
+
+INSERT into reviews Values(1, 11, 20, 5, 6, 5, 'One Size', 'Yes');
+
+--add items to favorites / get all favorited items 
+
+INSERT into favorites Values(1,2);
+INSERT into favorites Values(1,30);
+INSERT into favorites Values(1,26);
+
+SELECT pname, ptype, productid, price FROM favorites NATURAL JOIN product WHERE accountid = 1;
 
 --delete account
 
 Delete from customer where accountid = 1;
 
---invoice history should still be there 
+--invoice history should still be found
 
-select ird.fname, ird.lname, ird.invoiceid from invoice_record_deleted_account ird;
-select * from invoice;
+select ird.fname, ird.lname, ird.invoiceid from invoice_record_deleted_account ird where fname = 'Isabella' and lname = 'Attisano'; 
+
 -------------------------------------------------------------------------------------------
 
 -- KAYLEIGH ACCOUNT 
 -- Create accounts with ID 6, 7
 INSERT INTO customer VALUES(seqID.nextVal, 'Kayleigh', 'DiNatale', 'kwd@villanova.edu', '9141234567');
+INSERT INTO customer VALUES(seqID.nextVal, 'Regan', 'DiNatale', 'rdin@ohiostate.edu', '9178342938');
 
 -- Add addresses to Kayleigh's account
 INSERT INTO address VALUES(2, '123 Green St', 'Philadelphia', 'PA', '10456');
@@ -73,37 +117,43 @@ FROM address
 WHERE accountID = 2;
 
 -- ADD payments to Kayleigh's account
-INSERT INTO payment VALUES(2, 2, 2028, '5555444433332222', '123');
-INSERT INTO payment VALUES(2, 1, 2021, '1234567891234567', '423');
+INSERT INTO payment VALUES(6, 2, 2028, '5555444433332222', '123');
+INSERT INTO payment VALUES(6, 1, 2021, '1234567891234567', '423');
 
 -- View Kayleigh's payment options
 select expMonth ||'/'|| expYear AS Expiration_Date, '************'||''||SUBSTR(cardnumber, 12, 4) AS Card_Number 
 FROM payment 
-WHERE accountID = 2;
+WHERE accountID = 6;
 
 -- Add items to cart
-INSERT INTO cart_items VALUES(2, 1, 'medium', 1, 19.95);
-INSERT INTO cart_items VALUES(2, 3, 'medium', 1, 49.95);
-INSERT INTO cart_items VALUES(2, 4, 'medium', 3, 25);
+INSERT INTO cart_items VALUES(6, 1, 'medium', 1, 19.95);
+INSERT INTO cart_items VALUES(6, 3, 'medium', 1, 49.95);
+INSERT INTO cart_items VALUES(6, 4, 'medium', 3, 25);
 
 -- View Kayleigh's current cart
 select * from cart_items where cartid = 2;
 
--- Change quantity
+--fix total price input 
+Update cart_items set price = 75 where cartid = 2 and productid = 4;
+
+-- Change quantity, change price
 Update cart_items set pquantity = 2 where cartid = 2 and productid = 4;
+Update cart_items set price = 50 where cartid = 2 and productid = 4;
+
+select * from cart_items where cartid = 2
 
 --Update kayleigh cart to cause purchase and store in invoice
-Update cart set cartid = 2 where cartid = 2;
+Update cart set cartid = 6 where cartid = 2;
 
 -- Check trigger worked for kayleigh's invoice
 select * from invoice where accountid = 2
 
 --Check invoice products purchased
-select * from invoice_products where invoiceid = 3
+select * from invoice_products where invoiceid = 1
 
 select * from product
 --Create a review for Kayleigh
-INSERT into reviews Values(2, 14, 20, 5, 9, 4, 'True to Size', 'Yes');
+INSERT into reviews Values(2, 2, 20, 5, 9, 4, 'True to Size', 'Yes');
 
 --Check Kayleigh's reviews
 select age, feet ||' foot '|| inches as height, rating ||''|| ' out of 5' as Overall_Rating, sizedesc, recommend from reviews where accountid = 2;
@@ -111,21 +161,16 @@ select age, feet ||' foot '|| inches as height, rating ||''|| ' out of 5' as Ove
 --Check all reviews for a product
 select age, feet ||' foot '|| inches as height, rating ||''|| ' out of 5' as Overall_Rating, sizedesc, recommend from reviews where productid = 14;
 
+--Select statement to show items in Kayleigh's Cart
+select Pname, ProductID Psize, Pquantity, Price from cart_items NATURAL JOIN product where cartID = 2;
+
 --Cart total price, number of items, number of different types of products
 select SUM(Price) as Total_Price, SUM(pquantity) as Total_Quantity_of_Items, COUNT( DISTINCT ProductID) as Product_Variety 
 FROM cart_items NATURAL JOIN product 
 where CartID = 2;
 
---Add to cart
-INSERT into cart_items Values(2, 10, 'onesize', 4, 12.99);
-
---Add Review
-select * from cart_items
-INSERT into Reviews Values(2, 10, 20, 5, 9, 1, 'One Size', 'No');
-INSERT into Reviews Values(2, 42, 20, 5, 9, 5, 'One Size', 'Yes');
-
 -- Remove item from Kayleigh's cart
-DELETE FROM cart_items where cartid = 2 AND productID = 10;
+DELETE FROM cart_items where cartid = 2 AND productID = 3;
 
 -- Add to favorites
 INSERT INTO favorites VALUES(2, 2);
@@ -146,58 +191,46 @@ delete from customer where accountid = 2
 --CAITLIN ACCOUNT 
 
 --create account 
-select * from customer
+
 INSERT into customer VALUES(seqID.nextVal, 'Caitlin', 'van Goeverden', 'cvg@gmail.com', '4562348908');
-insert into cart_items values(3, 15, 'large', 3);
-delete from customer where accountid = 3;
+insert into cart_items values(142, 15, 'large', 3);
+delete from customer where accountid = 142;
 --add address / payment 
 INSERT INTO address(accountid, street, city, state, zip) VALUES('121', 'Lancaster', 'Villanova', 'PA', '19085');
 INSERT INTO payment(accountid, expmonth, expyear, cardnumber, securitycode) VALUES('121', '02', '23', '1234567890123456', '1234');
 
 -------------------------------------------------------------------------------------------
---RANDOM ACCOUNTS
--- Random Accounts 
+--Additional Accounts
 
-INSERT into customer VALUES(seqID.nextVal, 'customer', 'one', 'customerone@gmail.com', '0123456789');
-INSERT into customer VALUES(seqID.nextVal, 'customer', 'two', 'customertwo@gmail.com', '0123456789');
 INSERT into customer VALUES(seqID.nextVal, 'Mary', 'Smith', 'Mary234@icloud.com', '1923347659');
-INSERT INTO customer VALUES(seqID.nextVal, 'Regan', 'DiNatale', 'rdin@ohiostate.edu', '9178342938');
 
 --Sarah's Account
 INSERT into customer VALUES(seqID.nextVal, 'Sarah', 'Jones', 'Sjones@me.com', '3274828342');
 
---Sarah's cards
-INSERT into payment Values(8, 4, 2025, '1111222233334444', '012');
-INSERT into payment Values(8, 9, 2032, '1234123412341234', '914');
+INSERT into payment Values(5, 4, 2025, '1111222233334444', '012');
+INSERT into payment Values(5, 9, 2032, '1234123412341234', '914');
 
 --View Sarah's payment options
 select expMonth ||'/'|| expYear AS Expiration_Date, '************'||''||SUBSTR(cardnumber, 12, 4) AS Card_Number 
 FROM payment 
-WHERE accountID = 8;
+WHERE accountID = 5;
 
 --View How many forms of payment Sarah has
-select count(cardnumber) as Number_of_payments from payment where accountid = 8
+select count(cardnumber) as Number_of_payments from payment where accountid = 5
 
 --Sarah's addresses
-INSERT into address Values(8, '398 Brick Road', 'Tampa', 'FL', '25013');
-INSERT into address Values(8, '511 Greenwich Ave', 'New York', 'NY', '82315');
+INSERT into address Values(5, '398 Brick Road', 'Tampa', 'FL', '25013');
+INSERT into address Values(5, '511 Greenwich Ave', 'New York', 'NY', '82315');
 
 -- View Sarah's addresses
 SELECT Street ||' '|| City ||' '|| State ||' '|| Zip as Addresses 
 FROM address 
-WHERE accountID = 8;
+WHERE accountID = 5;
 
 -- Add items to cart 
-INSERT into cart_items Values(8, 6, 'small', 1, 14.95);
-INSERT into cart_items Values(8, 1, 'medium', 2, 19.95);
+INSERT into cart_items Values(5, 6, 'small', 1, 14.95);
+INSERT into cart_items Values(5, 1, 'medium', 2, 39.90);
 
 -- View cart items
-select * from cart_items where cartid = 8;
-
--- Purchase cart items 
-update cart set cartid = 8 where cartid = 8;
-
---View invoice
-select * from invoice where accountid = 8
-select * from invoice_products where invoiceid = 5
+select * from cart_items where cartid = 4;
